requirements.txt
streamlit
pandas
psycopg2-binary
sqlalchemy
python-dotenv
supabase
plotly
